visible map reduction counter added
added a counter that shows how many seconds there is left before the map is reduced. removed the draw_all function since it was becoming redundant (it would have too many params with those added to the draw method of top bar). now the shrink_counter is also decrementing instead of incrementing.