fixing a bug and introducing some optimization
fixed a bug where the crashed_players list was always resetting and thus the amount of lines in a game grew with great speed when ais crashed. this caused some big performance problems but is now fixed. soon to implement an optimization where the ais just check the lines that are near them